Prepare one step lookahead player for merge to master
